Add files via upload
Implementing the parallel computing in R is quite easy, although we know it should avoid to using loop in R.
In the case, we have to loop in large data sets, the parallel computing can save a lot of time.

For nested parallel computing, there is always  a question of which loop to parallelize. The standard advice is to parallelize the outer loop. However, if the outer loop doesn't have many iterations, programmer can parallelize the inner loop instead.

Another thing in R is there is  a nest loop operator, "%:%". This operator can change two loop into one. But it is not prefect right now. I mean sometimes the program still runs in sequentially instead of in parallel way.